+++ src/Core/Services/IOrderBookInitializer.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Lykke.Job.OrderBooksCacheProvider.Core.Services
+{
+    public interface IOrderBookInitializer
+    {
+        Task InitOrderBooks();
+    }
+}
