+++ src/Core/Services/IOrderBooksHandler.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Lykke.Job.OrderBooksCacheProvider.Core.Domain;
+
+namespace Lykke.Job.OrderBooksCacheProvider.Core.Services
+{
+    public interface IOrderBooksHandler
+    {
+        Task HandleOrderBook(OrderBook orderBook);
+    }
+}
