--- src/Lykke.Service.OrderBooksCacheProvider/appsettings.json
@@ -1,5 +1,0 @@
-﻿{
-  "ConnectionStrings": {
-    "Main": ""
-  }
-}
